Restrict permissions of who can delete an item

+++ sick-fits/backend/src/resolvers/Mutation.js
@@ -46,7 +46,16 @@ const Mutation = {
   async deleteItem(parent, args, ctx, info) {
     const where = { id: args.id };
 
-    const item = await ctx.db.query.item({ where }, `{id title}`);
+    const item = await ctx.db.query.item({ where }, `{ id title user { id } }`);
+
+    const ownsItem = item.user.id === ctx.request.userId;
+    const hasPermissions = ctx.request.user.permissions.some(permission =>
+      ["ADMIN", "ITEMDELETE"].includes(permission)
+    );
+
+    if (!ownsItem || !hasPermissions) {
+      throw new Error("You don't have permission to do that")
+    }
 
     return ctx.db.mutation.deleteItem({ where }, info);
   },
